dashboard(settings): Tor/Reaper cross-links, shared capability toggles, scheduled-backups mirror + type scale
